requirements.txt
torch==2.2.1+cu121
torchvision==0.17.1+cu121
transformers==4.40.0
diffusers==0.27.2
gradio==4.25.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.4
accelerate==0.28.0
--extra-index-url https://download.pytorch.org/whl/cu121